shared view model between activity and fragments

+++ app/src/main/java/com/example/noaa/model/Current.kt
@@ -10,7 +10,7 @@ data class Current(
     val sunrise: Int,
     val sunset: Int,
     val temp: Double,
-    val uvi: Int,
+    val uvi: Double,
     val visibility: Int,
     val weather: List<Weather>,
     val wind_deg: Int,
